+++ editors/code/src/extension.ts
@@ -19,7 +19,7 @@ import * as utils from "./utils";
 export type PlatformContext = { client: LanguageClient | undefined };
 export type AddArgCallback = (
   argument: string,
-  mode?: "All" | "Prod" | "Debug",
+  mode?: "All" | "Prod" | "Debug"
 ) => void;
 
 let client: LanguageClient | undefined = undefined;
@@ -43,7 +43,7 @@ const getFFlags = async () => {
     () =>
       fetch(CURRENT_FFLAGS)
         .then((r) => r.json() as Promise<FFlagsEndpoint>)
-        .then((r) => r.applicationSettings),
+        .then((r) => r.applicationSettings)
   );
 };
 
@@ -87,7 +87,7 @@ const startLanguageServer = async (context: vscode.ExtensionContext) => {
       if (vscode.workspace.workspaceFolders) {
         uri = utils.resolveUri(
           vscode.workspace.workspaceFolders[0].uri,
-          definitionPath,
+          definitionPath
         );
       } else {
         uri = vscode.Uri.file(definitionPath);
@@ -96,7 +96,7 @@ const startLanguageServer = async (context: vscode.ExtensionContext) => {
         addArg(`--definitions=${uri.fsPath}`);
       } else {
         vscode.window.showWarningMessage(
-          `Definitions file at ${definitionPath} does not exist, types will not be provided from this file`,
+          `Definitions file at ${definitionPath} does not exist, types will not be provided from this file`
         );
       }
     }
@@ -111,7 +111,7 @@ const startLanguageServer = async (context: vscode.ExtensionContext) => {
       if (vscode.workspace.workspaceFolders) {
         uri = utils.resolveUri(
           vscode.workspace.workspaceFolders[0].uri,
-          documentationPath,
+          documentationPath
         );
       } else {
         uri = vscode.Uri.file(documentationPath);
@@ -120,7 +120,7 @@ const startLanguageServer = async (context: vscode.ExtensionContext) => {
         addArg(`--docs=${uri.fsPath}`);
       } else {
         vscode.window.showWarningMessage(
-          `Documentations file at ${documentationPath} does not exist`,
+          `Documentations file at ${documentationPath} does not exist`
         );
       }
     }
@@ -150,7 +150,7 @@ const startLanguageServer = async (context: vscode.ExtensionContext) => {
       }
     } catch (err) {
       vscode.window.showWarningMessage(
-        "Failed to fetch current Luau FFlags: " + err,
+        "Failed to fetch current Luau FFlags: " + err
       );
     }
   }
@@ -168,7 +168,7 @@ const startLanguageServer = async (context: vscode.ExtensionContext) => {
     for (let [name, value] of Object.entries(overridenFFlags)) {
       if (!isAlphanumericUnderscore(name)) {
         vscode.window.showWarningMessage(
-          `Invalid FFlag name: '${name}'. It can only contain alphanumeric characters`,
+          `Invalid FFlag name: '${name}'. It can only contain alphanumeric characters`
         );
       }
 
@@ -185,7 +185,7 @@ const startLanguageServer = async (context: vscode.ExtensionContext) => {
     command: vscode.Uri.joinPath(
       context.extensionUri,
       "bin",
-      os.platform() === "win32" ? "server.exe" : "server",
+      os.platform() === "win32" ? "server.exe" : "server"
     ).fsPath,
     args,
   };
@@ -196,7 +196,7 @@ const startLanguageServer = async (context: vscode.ExtensionContext) => {
       context.extensionUri,
       "..",
       "..",
-      process.env["SERVER_PATH"] ?? "unknown.exe",
+      process.env["SERVER_PATH"] ?? "unknown.exe"
     ).fsPath,
     args: debugArgs,
   };
@@ -211,8 +211,12 @@ const startLanguageServer = async (context: vscode.ExtensionContext) => {
       { language: "luau", scheme: "untitled" },
     ],
     diagnosticPullOptions: {
-      onChange: true,
-      onSave: true,
+      onChange: vscode.workspace
+        .getConfiguration("luau-lsp.diagnostics")
+        .get("pullOnChange", true),
+      onSave: vscode.workspace
+        .getConfiguration("luau-lsp.diagnostics")
+        .get("pullOnSave", true),
     },
     initializationOptions: {
       fflags,
@@ -223,7 +227,7 @@ const startLanguageServer = async (context: vscode.ExtensionContext) => {
     "luau",
     "Luau Language Server",
     serverOptions,
-    clientOptions,
+    clientOptions
   );
 
   platformContext.client = client;
@@ -254,7 +258,7 @@ export async function activate(context: vscode.ExtensionContext) {
       if (client) {
         client.sendNotification("$/flushTimeTrace");
       }
-    }),
+    })
   );
 
   context.subscriptions.push(
@@ -263,7 +267,7 @@ export async function activate(context: vscode.ExtensionContext) {
         vscode.window
           .showInformationMessage(
             "Luau FFlags have been changed, reload server for this to take effect.",
-            "Reload Language Server",
+            "Reload Language Server"
           )
           .then((command) => {
             if (command === "Reload Language Server") {
@@ -277,7 +281,7 @@ export async function activate(context: vscode.ExtensionContext) {
         vscode.window
           .showInformationMessage(
             "Luau type definitions have been changed, reload server for this to take effect.",
-            "Reload Language Server",
+            "Reload Language Server"
           )
           .then((command) => {
             if (command === "Reload Language Server") {
@@ -285,7 +289,7 @@ export async function activate(context: vscode.ExtensionContext) {
             }
           });
       }
-    }),
+    })
   );
 
   await startLanguageServer(context);